// packages/unit/chat/webpack/types/types.ts
export interface BuildPath {
    entry: string;
    html: string;
    public: string;
    output: string;
    src: string;
}

export type BuildMode = 'production' | 'development';
export type BuildPlatform = 'mobile' | 'desktop';

export interface BuildOptions {
    port: number;
    paths: BuildPath;
    mode: BuildMode;
    analyzer?: boolean;
    platform?: BuildPlatform
}